% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{cut_short_scale_german}
\alias{cut_short_scale_german}
\title{Short scale German}
\usage{
cut_short_scale_german(space = TRUE)
}
\arguments{
\item{space}{logical whether to add a space before the abbreviation}
}
\value{
named numeric vector
}
\description{
Return a named numeric vector with German short scale abbreviations.
}
