% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_centroids}
\alias{get_centroids}
\title{Get Centroids of Spatial Features}
\usage{
get_centroids(sf_object)
}
\arguments{
\item{sf_object}{A spatial object of class sf, sfc or sfg.}
}
\value{
A tibble with the coordinates of the centroids (X, Y) of the spatial features in the input object.
}
\description{
This function calculates the centroids of the spatial features in a given spatial object.
The centroids are calculated as points on the surface of the spatial features.
}
\examples{
\dontrun{
sf_object <- sf::st_read('path_to_your_shapefile.shp')
centroids <- get_centroids(sf_object)
}
}
