% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scatterpie}
\alias{plot_scatterpie}
\title{Plot Scatterpie Chart}
\usage{
plot_scatterpie(
  sf_object,
  var,
  cols = NULL,
  legend_name = "Typ",
  ratio = 0.08,
  area = TRUE,
  legend.x = 142767.4,
  legend.y = 483225.9,
  legend.n = 3,
  color.pie = NA,
  alpha.pie = 0.9,
  background_plot = NULL,
  ...
)
}
\arguments{
\item{sf_object}{A spatial object of class sf, sfc or sfg.}

\item{var}{A string representing the name of the variable in the data frame to be transformed.}

\item{cols}{A vector of variables to be used for the pie slices in the scatterpie plot. If NULL only \code{var} will be used (default is NULL).}

\item{legend_name}{A string to be used as the legend title (default is "Typ").}

\item{ratio}{A numeric value representing the ratio used for scaling (default is 0.08).}

\item{area}{A logical value indicating whether the transformation should be based on area (default is TRUE).}

\item{legend.x}{A numeric value representing the x-coordinate of the legend.}

\item{legend.y}{A numeric value representing the y-coordinate of the legend.}

\item{legend.n}{An integer representing the number of circles in the legend.}

\item{color.pie}{The line color of the pie slices in the scatterpie plot (default is NA).}

\item{alpha.pie}{The transparency of the pie slices in the scatterpie plot (default is 0.9).}

\item{background_plot}{A ggplot object to be used as the background of the scatterpie plot (default is NULL.)}

\item{...}{Additional arguments to be passed to the \code{geom_scatterpie_legend} function.}
}
\value{
A ggplot2 scatterpie plot.
}
\description{
This function creates a scatterpie plot using ggplot2. It overlays scatterpie points on a spatial plot,
where the size of the scatterpie points are scaled to \code{var} and the pie slices are scaled to \code{cols}.
The size of the scatterpie points can be scaled by \code{area} or radius.
}
\examples{
\dontrun{
sf_object <- sf::st_read("path_to_your_shapefile.shp")
plot <- plot_scatterpie(sf_object, "your_variable")
}
}
