% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_sf}
\alias{transform_sf}
\title{Transform Spatial Data}
\usage{
transform_sf(sf_object, var, ratio = 0.08, area = TRUE)
}
\arguments{
\item{sf_object}{A spatial object of class sf, sfc or sfg.}

\item{var}{A string representing the name of the variable in the spatial object to be transformed.}

\item{ratio}{A numeric value representing the ratio used for scaling (default is 0.08).}

\item{area}{A logical value indicating whether the transformation should be based on area (default is TRUE).}
}
\value{
A list containing the data frame with the transformed variable and the multiplier used for the transformation.
}
\description{
This function calculates the centroids of the spatial features in a given spatial object and
transforms a variable based on the area of spatial features.
It then adds the transformed variable as a new column (r) to the data frame.
}
\details{
Convenience function that wraps \code{\link{get_centroids}} and \code{\link{transform_data}}.
It returns a data frame, since \pkg{scatterpie} uses data frames as input.
}
\examples{
\dontrun{
sf_object <- sf::st_read('path_to_your_shapefile.shp')
transformed_data <- transform_data(data, 'your_variable', sf_object, ratio = 0.1, area = TRUE)
}
}
