% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_values}
\alias{transform_values}
\title{Transforms \code{values} for \pkg{scatterpie}}
\usage{
transform_values(
  values,
  sf_object,
  ratio = 0.08,
  area = FALSE,
  return_multiplier = FALSE
)
}
\arguments{
\item{values}{A numeric vector representing the values/size of the spatial data points.}

\item{sf_object}{A spatial data object in accordance with which the values are transformed.}

\item{ratio}{A numeric value representing the ratio used for scaling the values (default is 0.08).}

\item{area}{A logical value indicating whether the transformation should be based on the area instead of the radius (default is FALSE).}

\item{return_multiplier}{A logical value indicating whether to return the multiplier used for the transformation (default is FALSE).}
}
\value{
Returns the transformed values as a vector.
If return_multiplier is TRUE, the function returns the transformed values and the multiplier used for the transformation as a list.
}
\description{
\code{transform_values} transforms the given \code{values} in accordance with a spatial data object so it can be used in the \pkg{scatterpie} package.
The values can optionally be transformed so they match the \code{area} of the plotted circles instead of the radius and scales it by a specified \code{ratio.}
It can also return the \code{multiplier} used for the transformation instead of the transformed values.
}
\examples{
\dontrun{
values <- c(100, 10, 200, 40)
sf_object <- sf::st_read('path_to_your_shapefile.shp')
r <- transform_values(
  values = values,
  sf_object = sf_object,
  ratio = 0.08,
  area = TRUE,
  return_multiplier = FALSE
)
}
}
